# openapi.yaml
openapi: 3.0.3
info:
  title: LeadPhone Validator API
  version: 1.1.0
  description: High performance microservice in Go to validate phone numbers, decode DDD and DDI codes, normalize Brazilian 9th digits, detect suspicious patterns, and generate WhatsApp links.

servers:
  - url: http://localhost:3007
    description: Local development server

paths:
  /:
    get:
      summary: Health check endpoint
      description: Returns service availability status and version.
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /health:
    get:
      summary: Health check endpoint alias
      description: Returns service availability status and version.
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /v1/validate:
    post:
      summary: Validate and format single phone number
      description: Parses single phone number, extracts carrier and geocoding metadata, and generates formatted representations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateRequest"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResponse"

  /v1/validate/batch:
    post:
      summary: Validate multiple phone numbers in batch
      description: Concurrently validates an array of phone numbers using worker goroutines.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchValidateRequest"
      responses:
        "200":
          description: Batch validation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchValidationResponse"

  /v1/ddd:
    get:
      summary: Brazilian Area Code (DDD) Lookup
      description: Look up Brazilian area code details, state, region, timezone, and major covered municipalities.
      parameters:
        - name: q
          in: query
          description: Area code (e.g. 11, 48) or municipality name (e.g. Curitiba)
          schema:
            type: string
      responses:
        "200":
          description: DDD lookup results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DDDResponse"

  /v1/ddi:
    get:
      summary: International Calling Code (DDI) Lookup
      description: Look up international country calling codes, ISO codes, exit prefixes, and emergency numbers.
      parameters:
        - name: q
          in: query
          description: Calling code (e.g. 55, 65) or country name (e.g. Singapore)
          schema:
            type: string
      responses:
        "200":
          description: DDI lookup results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DDIResponse"

  /v1/sanitize/ninth-digit:
    post:
      summary: Brazilian 9th Digit Normalizer
      description: Automatically converts legacy 8-digit mobile numbers to 9 digits while preserving 8-digit landlines.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NinthDigitRequest"
      responses:
        "200":
          description: Ninth digit normalization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NinthDigitResponse"

  /v1/whatsapp/link:
    post:
      summary: WhatsApp Link Builder
      description: Generates clean wa.me direct links with optional URL-encoded greeting message templates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsAppLinkRequest"
      responses:
        "200":
          description: WhatsApp link result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsAppLinkResponse"

  /v1/check/patterns:
    post:
      summary: Suspicious Pattern and Dummy Detector
      description: Inspects phone numbers for repeated digits, sequential ranges, and fictitious test numbers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatternCheckRequest"
      responses:
        "200":
          description: Pattern analysis report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatternCheckResponse"

components:
  schemas:
    ValidateRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          example: "+55 (11) 98765-4321"
        country:
          type: string
          example: "BR"

    BatchValidateRequest:
      type: object
      required:
        - phones
      properties:
        phones:
          type: array
          items:
            $ref: "#/components/schemas/ValidateRequest"

    ValidationResponse:
      type: object
      properties:
        input:
          type: object
        parsed:
          type: object
        metadata:
          type: object
        formatted:
          type: object

    BatchValidationResponse:
      type: object
      properties:
        total:
          type: integer
        valid:
          type: integer
        invalid:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/ValidationResponse"

    DDDResponse:
      type: object
      properties:
        query:
          type: string
        found:
          type: boolean
        results:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              state:
                type: string
              stateName:
                type: string
              region:
                type: string
              majorCities:
                type: array
                items:
                  type: string
              timezone:
                type: string

    DDIResponse:
      type: object
      properties:
        query:
          type: string
        found:
          type: boolean
        results:
          type: array
          items:
            type: object
            properties:
              callingCode:
                type: string
              countryName:
                type: string
              iso2:
                type: string
              iso3:
                type: string
              iddPrefix:
                type: string
              emergencyNumbers:
                type: array
                items:
                  type: string
              timezoneHint:
                type: string

    NinthDigitRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          example: "1187654321"

    NinthDigitResponse:
      type: object
      properties:
        input:
          type: string
        normalized:
          type: string
        changed:
          type: boolean
        isMobile:
          type: boolean
        isLandline:
          type: boolean
        isValid:
          type: boolean
        message:
          type: string

    WhatsAppLinkRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          example: "+55 11 98765-4321"
        message:
          type: string
          example: "Hello, scheduling demo"

    WhatsAppLinkResponse:
      type: object
      properties:
        phone:
          type: string
        e164:
          type: string
        message:
          type: string
        url:
          type: string
        isValid:
          type: boolean

    PatternCheckRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          example: "11999999999"

    PatternCheckResponse:
      type: object
      properties:
        phone:
          type: string
        isSuspicious:
          type: boolean
        flags:
          type: array
          items:
            type: string
        score:
          type: integer
        description:
          type: string

    HealthResponse:
      type: object
      properties:
        status:
          type: string
        service:
          type: string
        version:
          type: string